requirements.txt
ioc-finder
phonenumbers
quantulum3
nltk
geograpy3
pycountry
pytest
pytest-assume
git+https://github.com/fluquid/find_job_titles.git